Updated Signup Page
fixes https://linear.app/ghost/issue/DES-881/sign-up-flow-on-mobile-with-checkbox-on-requires-ux-improvement

On mobile, the Sign Up button is no longer sticky. This ensures customers scroll down and see the checkbox before they hit the button.

If they do still miss the checkbox, we now have a check in place that — if the checkbox is the last missing thing — scrolls the checkbox into view upon submission.